refactor(questions): change folder structure
