+++ client/app/bundles/course/assessment/question/voice-responses/commons/validation.ts
@@ -1,0 +1,5 @@
+import { qnFormCommonFieldsValidation } from '../../components/QuestionFormCommonFields';
+
+const questionSchema = qnFormCommonFieldsValidation;
+
+export default questionSchema;
