--- client/app/bundles/course/assessment/question/voice-responses/validation.ts
@@ -1,5 +1,0 @@
-import { qnFormCommonFieldsValidation } from '../components/QuestionFormCommonFields';
-
-const questionSchema = qnFormCommonFieldsValidation;
-
-export default questionSchema;
